requirements.txt
Flask==3.1.0
flask-mysqldb==2.0.0
Flask-Mail==0.10.0
Werkzeug==3.1.3
openpyxl==3.1.5
python-dotenv==1.0.1
